Implement Edit functionality, and difference edit scenarios regarding the images

+++ src/services/apiCabins.js
@@ -11,33 +11,47 @@ export async function getCabins() {
   return data;
 }
 
-export async function insertCabin(newCabin) {
+export async function createEditCabin(newCabin, id = null) {
+  const hasImagePath = newCabin?.image?.startsWith?.(supabaseUrl);
+  //Construct Image path
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  //Create/Edit cabin
+  let query = supabase.from("cabins");
+
+  //A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  //B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  //C) Run query
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error("Could not add new cabin.");
   }
 
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    //Upload image
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-    console.error(storageError);
-
-    throw new Error(
-      "Cabin image could not be uploaded. Cabin creation canceled."
-    );
+    //Delete the cabin if there was an error uploading the image
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      throw new Error(
+        "Cabin image could not be uploaded. Cabin creation canceled."
+      );
+    }
   }
 
   return data;
